+++ src/CPPMoMinterface/content/CPPMoMsolver/src/triangle.cpp
@@ -14,7 +14,7 @@
  *
  */
 
-Triangle::Triangle(int vertex_1, int vertex_2, int vertex_3, Node centre, float area)
+Triangle::Triangle(int vertex_1, int vertex_2, int vertex_3, Node centre, double area)
 {
     this->vertex_1 = vertex_1;  
     this->vertex_2 = vertex_2;  
@@ -66,7 +66,7 @@ std::vector<int> Triangle::getEdges()
     return this->edge_indices;
 }
 
-float Triangle::getArea()
+double Triangle::getArea()
 {
     return this->area;
 }
@@ -81,7 +81,7 @@ void Triangle::setEdgeIndex(int index)
     this->edge_indices.push_back(index);
 }
 
-float Triangle::calculateArea(Node node_vertex_1, Node node_vertex_2, Node node_vertex_3)
+double Triangle::calculateArea(Node node_vertex_1, Node node_vertex_2, Node node_vertex_3)
 {
     // Lets calculate the area
     // Easiest way is to use Heron's formula
@@ -89,12 +89,12 @@ float Triangle::calculateArea(Node node_vertex_1, Node node_vertex_2, Node node_
     // Area = sqrt(S(S-A)(S-B)(S-C))
 
     // First lets get A, B and C
-    float a = node_vertex_1.getDistanceTo(node_vertex_2);
-    float b = node_vertex_1.getDistanceTo(node_vertex_3);
-    float c = node_vertex_2.getDistanceTo(node_vertex_3);
+    double a = node_vertex_1.getDistanceTo(node_vertex_2);
+    double b = node_vertex_1.getDistanceTo(node_vertex_3);
+    double c = node_vertex_2.getDistanceTo(node_vertex_3);
     
     // Then lets get S
-    float s = (a + b + c) / 2;
+    double s = (a + b + c) / 2;
 
     // Finally, lets get the area
     return std::sqrt(s * (s - a) * (s - b) * (s - c));  
@@ -105,15 +105,15 @@ Node Triangle::calculateCentre(Node node_vertex_1, Node node_vertex_2, Node node
     // Lets calculate the centre
     // The centre is the average of the three nodes
 
-    float x_coord = (node_vertex_1.getXCoord() + 
+    double x_coord = (node_vertex_1.getXCoord() + 
                      node_vertex_2.getXCoord() + 
                      node_vertex_3.getXCoord()) / 3;
 
-    float y_coord = (node_vertex_1.getYCoord() + 
+    double y_coord = (node_vertex_1.getYCoord() + 
                      node_vertex_2.getYCoord() + 
                      node_vertex_3.getYCoord()) / 3;
 
-    float z_coord = (node_vertex_1.getZCoord() + 
+    double z_coord = (node_vertex_1.getZCoord() + 
                      node_vertex_2.getZCoord() + 
                      node_vertex_3.getZCoord()) / 3;
